backend/requirements.txt
Django>=4.2,<5
djangorestframework>=3.14
django-cors-headers>=4.4
python-dotenv>=1.0
pdfplumber>=0.11
PyPDF2>=3.0
uvicorn[standard]>=0.30
ollama>=0.3.0
